Move TypeScript water monitoring files to appropriate client dir, add back webpack config, remove/fix test files
