+++ WNPRC_EHR/webpack.config.js
@@ -29,40 +29,6 @@ let breedingConfig = function wp(env) {
     };
 };
 
-let testConfig = function wp(env) {
-
-    return {
-        devtool: 'source-map',
-        entry: './src/client/test.tsx',
-        externals: {
-            jquery: 'jQuery',
-            urijs: 'URI',
-        },
-        mode: process.env.NODE_ENV,
-        module: {
-            rules: [
-                {
-                    test: /\.tsx?$/,
-                    loader: "ts-loader"
-                },
-                {
-                    test: /\.css$/,
-                    loaders: ['style-loader', 'css-loader', 'sass-loader']
-                }
-            ],
-        },
-        output: {
-            filename: 'test.js',
-            library: 'Test',
-            libraryExport: 'default',
-            libraryTarget: 'umd',
-            path: path.resolve(__dirname, 'resources/web/wnprc_ehr/gen')
-        },
-        resolve: {
-            extensions: ['.ts', '.tsx', '.js', '.json'],
-        },
-    };
-};
 
 let feedingConfig = function wp(env) {
 
@@ -196,6 +162,41 @@ var weightConfig = function wp(env) {
     };
 };
 
+let waterMonitoringConfig = function wp(env) {
+
+    return {
+        devtool: 'source-map',
+        entry: './src/client/watermonitoring/waterMonitoringSystem.ts',
+        externals: {
+            jquery: 'jQuery',
+            urijs: 'URI',
+            LABKEY: 'LABKEY'
+        },
+        mode: process.env.NODE_ENV,
+        module: {
+            rules: [
+                {
+                    loader: 'ts-loader',
+                    test: /\.tsx?$/},
+                {
+                    test: /\.css$/i,
+                    use: ['style-loader', 'css-loader']
+                },
+            ],
+        },
+        output: {
+            filename: 'waterMonitoringSystem.js',
+            library: 'waterMonitoringSystem',
+            libraryExport: 'default',
+            libraryTarget: 'umd',
+            path: path.resolve(__dirname, 'resources/web/wnprc_ehr/gen')
+        },
+        resolve: {
+            extensions: ['.ts', '.tsx', '.js', '.json', '.css'],
+        },
+    };
+};
+
 module.exports = [
-    breedingConfig, testConfig, feedingConfig, researchUltrasoundsConfig, abstractConfig, weightConfig
+    breedingConfig, feedingConfig, researchUltrasoundsConfig, abstractConfig, weightConfig, waterMonitoringConfig
 ];
